requirements.txt
# Passkeys CLI - Production Dependencies
# Python 3.8+

# Database
psycopg[binary]>=3.1.0

# Cryptography
cryptography>=41.0.0

# Password Hashing
argon2-cffi>=23.1.0

# Environment Variables
python-dotenv>=1.0.0
